+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Release available for download on [GitHub](https://github.com/microsoft/DirectX-
 
 ### November 2024
 * Converted all samples to use VS 2019 instead of VS 2008 or VS 2010
-* C++11 modernization for some Direct3D 11 samples using GitHub releases of DXUT and FX11
+* C++11 modernization for some Direct3D 11 samples using GitHub releases of DXUT
+* C++11 modernization and porting some samples to Effects for Direct3D 11 using GitHub release of FX11
 * Removed use of legacy DirectX SDK and instead uses Windows SDK and the `Microsoft.DXSDK.D3DX` NuGet package
 * Removed XACT samples
 * Added legacy DirectSound samples